Geometries: move specific geometry files into their own subdirectory.
