+++ source/geometries/geometry_disk.h
