+++ source/geometries/geometry_geotiff_profile.h
